adds network and EVM routes
