+++ gateway/src/connectors/pangolin/pangolin.ts
@@ -17,8 +17,8 @@ import {
   Trade,
 } from '@pangolindex/sdk';
 import { logger } from '../../services/logger';
-import { ExpectedTrade, Uniswapish } from '../../services/uniswapish.interface';
 import { Avalanche } from '../../chains/avalanche/avalanche';
+import { ExpectedTrade, Uniswapish } from '../../services/common-interfaces';
 
 export class Pangolin implements Uniswapish {
   private static _instances: { [name: string]: Pangolin };
