+++ gateway/src/network/network.requests.ts
@@ -1,0 +1,32 @@
+import {
+  CustomTransactionReceipt,
+  CustomTransactionResponse,
+  NetworkSelectionRequest,
+} from '../services/common-interfaces';
+
+export interface BalanceRequest extends NetworkSelectionRequest {
+  address: string; // the users public Ethereum key
+  tokenSymbols: string[]; // a list of token symbol
+}
+
+export interface BalanceResponse {
+  network: string;
+  timestamp: number;
+  latency: number;
+  balances: Record<string, string>; // the balance should be a string encoded number
+}
+
+export interface PollRequest extends NetworkSelectionRequest {
+  txHash: string;
+}
+
+export interface PollResponse {
+  network: string;
+  timestamp: number;
+  currentBlock: number;
+  txHash: string;
+  txStatus: number;
+  txBlock: number;
+  txData: CustomTransactionResponse | null;
+  txReceipt: CustomTransactionReceipt | null;
+}
